Change entities ids from integer to long. Use identity strategy for id generation

+++ billy-core/src/main/java/com/premiumminds/billy/core/services/entities/Entity.java
@@ -31,7 +31,7 @@ import com.premiumminds.billy.core.services.UID;
  */
 public interface Entity extends Serializable {
 
-    public Integer getID();
+    public Long getID();
 
     /**
      * Gets the entity unique identifier.
